fix: rename StopServiceCommand name typo
